# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hayyah CRM API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: customers
    description: Customer management
  - name: bookings
    description: Booking management
  - name: technicians
    description: Technician management
  - name: users
    description: User directory (hayyah v1)
  - name: services
    description: Service catalog
  - name: dashboard
    description: Dashboard metrics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, lead]
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomer"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomer"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, in_progress, completed, cancelled]
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBooking"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    put:
      operationId: updateBooking
      tags: [bookings]
      summary: Update a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBooking"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /v1/user/create:
    post:
      operationId: createUser
      tags: [users]
      summary: Create app user (Spring POST /api/v1/user/create)
      description: >
        Proxies to Hayyah `AppUserController`: new user with no pre-existing Keycloak id.
        Same email validation and `mobileNumber` normalization as the backend; returns 201 + AppUserDto.
        Live schema: SpringDoc `/v3/api-docs` or Swagger UI when enabled.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppUser"
      responses:
        "201":
          description: Created — AppUserDto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUserDto"

  /v1/user/createExternal:
    post:
      operationId: createExternalAppUser
      tags: [users]
      summary: Create/sync external app user (Spring POST /api/v1/user/createExternal)
      description: >
        Keycloak "external" path via `createExternalAppUser`: user id is usually already known
        (Keycloak or another system); behavior if the id does not exist depends on Keycloak.
        Backend uses `permitAll()`. Proxied without requiring a Bearer token (optional if sent).
        Returns 201 + AppUserDto.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExternalAppUser"
      responses:
        "201":
          description: Created or synced — AppUserDto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUserDto"

  /v1/technicians/admin/{userId}:
    post:
      operationId: registerTechnicianByAdmin
      tags: [technicians]
      summary: Register technician profile by admin for an existing user
      description: >
        Admin-only endpoint: `POST /api/v1/technicians/admin/{userId}`.
        Promotes target user to provider when needed, then creates technician profile.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterTechnicianByAdmin"
      responses:
        "201":
          description: Technician profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"

  /v1/technicians:
    get:
      operationId: listTechnicians
      tags: [technicians]
      summary: List all technicians
      responses:
        "200":
          description: List of technicians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Technician"
    post:
      operationId: createTechnician
      tags: [technicians]
      summary: Register technician profile for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTechnician"
      responses:
        "201":
          description: Created technician
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"

  /dashboard/metrics:
    get:
      operationId: getDashboardMetrics
      tags: [dashboard]
      summary: Get dashboard overview metrics
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardMetrics"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, inactive, lead]
        totalBookings:
          type: integer
        totalSpent:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - phone
        - address
        - status
        - totalBookings
        - totalSpent
        - createdAt

    CreateCustomer:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, inactive, lead]
      required:
        - name
        - email
        - phone
        - address

    Booking:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        technicianId:
          type: integer
        technicianName:
          type: string
        serviceId:
          type: integer
        serviceName:
          type: string
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled]
        scheduledAt:
          type: string
        address:
          type: string
        notes:
          type: string
        price:
          type: number
        createdAt:
          type: string
      required:
        - id
        - customerId
        - customerName
        - serviceId
        - serviceName
        - status
        - scheduledAt
        - address
        - price
        - createdAt

    CreateBooking:
      type: object
      properties:
        customerId:
          type: integer
        technicianId:
          type: integer
        serviceId:
          type: integer
        scheduledAt:
          type: string
        address:
          type: string
        notes:
          type: string
        price:
          type: number
      required:
        - customerId
        - serviceId
        - scheduledAt
        - address
        - price

    UpdateBooking:
      type: object
      properties:
        technicianId:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled]
        scheduledAt:
          type: string
        notes:
          type: string
        price:
          type: number

    Technician:
      type: object
      description: Hayyah v1 technician profile (hayyah.me/api/v1/technicians)
      properties:
        id:
          type: string
        userId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        specialization:
          $ref: "#/components/schemas/TechnicianSpecialization"
        verified:
          type: boolean
        rating:
          type: number
          nullable: true
        bio:
          type: string
          nullable: true
        createdAt:
          type: integer
          format: int64
      required:
        - id
        - userId
        - firstName
        - lastName
        - email
        - specialization
        - verified
        - createdAt

    TechnicianSpecialization:
      type: string
      description: Case-insensitive on server; send uppercase.
      enum:
        - CLEANER
        - LAUNDRY
        - ELECTRICIAN
        - PAINTER
        - PLUMBER
        - CARPENTER
        - HANDYMAN
        - MOVER
        - CHEF
        - HVAC
        - LOCKSMITH

    CreateTechnician:
      type: object
      description: >
        User id comes from JWT (sub), not the body. Caller must be provider or admin.
      properties:
        specialization:
          $ref: "#/components/schemas/TechnicianSpecialization"
        bio:
          type: string
      required:
        - specialization

    CreateAppUser:
      type: object
      description: >
        Request body for POST /api/v1/user/create (new user). JSON property names match Spring (camelCase).
      properties:
        firstName:
          type: string
        lastName:
          type: string
        userName:
          type: string
        email:
          type: string
          format: email
        mobileNumber:
          type: string
          description: E.164-style after normalization (e.g. +971501234567)
        role:
          type: string
          description: Application role (e.g. user, provider)
      required:
        - firstName
        - lastName
        - userName
        - email
        - mobileNumber
        - role

    CreateExternalAppUser:
      type: object
      description: >
        Request body for POST /api/v1/user/createExternal — sync/profile when external user id is known.
      properties:
        id:
          type: string
          description: Keycloak or external system user id
        firstName:
          type: string
        lastName:
          type: string
        userName:
          type: string
        email:
          type: string
          format: email
        mobileNumber:
          type: string
        role:
          type: string
      required:
        - id
        - firstName
        - lastName
        - userName
        - email
        - mobileNumber
        - role

    AppUserDto:
      type: object
      description: >
        Spring response for user create endpoints (201). See AppUserController / SpringDoc for full schema.
      properties:
        id:
          type: string
        userName:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        mobileNumber:
          type: string
        role:
          type: string

    RegisterTechnicianByAdmin:
      type: object
      properties:
        specialization:
          $ref: "#/components/schemas/TechnicianSpecialization"
        bio:
          type: string
      required:
        - specialization

    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        basePrice:
          type: number
        durationMinutes:
          type: integer
      required:
        - id
        - name
        - description
        - category
        - basePrice
        - durationMinutes

    DashboardMetrics:
      type: object
      properties:
        totalCustomers:
          type: integer
        activeBookings:
          type: integer
        completedToday:
          type: integer
        revenue:
          type: number
        revenueGrowth:
          type: number
        newLeads:
          type: integer
        availableTechnicians:
          type: integer
        upcomingBookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        recentCustomers:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        bookingsByStatus:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              count:
                type: integer
            required:
              - status
              - count
      required:
        - totalCustomers
        - activeBookings
        - completedToday
        - revenue
        - revenueGrowth
        - newLeads
        - availableTechnicians
        - upcomingBookings
        - recentCustomers
        - bookingsByStatus