# lib/api-zod/src/generated/types/registerTechnicianByAdmin.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * Hayyah CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { TechnicianSpecialization } from "./technicianSpecialization";

export interface RegisterTechnicianByAdmin {
  specialization: TechnicianSpecialization;
  bio?: string;
}