# lib/api-zod/src/generated/types/technician.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * Hayyah CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { TechnicianSpecialization } from "./technicianSpecialization";

/**
 * Hayyah v1 technician profile (hayyah.me/api/v1/technicians)
 */
export interface Technician {
  id: string;
  userId: string;
  firstName: string;
  lastName: string;
  email: string;
  specialization: TechnicianSpecialization;
  verified: boolean;
  rating?: number | null;
  bio?: string | null;
  createdAt: number;
}